Refactor: 카테고리 랜덤 생성 함수 분리
