+++ src/App.js
@@ -1,4 +1,5 @@
 const { Console, Random } = require("@woowacourse/mission-utils");
+const { randomCategory } = require("./GetRandomCategories");
 
 const SAMPLE = {
   일식: "규동, 우동, 미소시루, 스시, 가츠동, 오니기리, 하이라이스, 라멘, 오코노미야끼",
@@ -17,19 +18,6 @@ const categoryKey = {
   5: "양식",
 };
 
-const getCategories = (categories) => {
-  while (true) {
-    const curCategory = Random.pickNumberInRange(1, 5);
-    const count = categories.filter(
-      (category) => category === curCategory,
-    ).length;
-
-    if (count > 1) continue;
-
-    return curCategory;
-  }
-};
-
 class App {
   #coaches; // 코치들이 저장되는 배열. 코치의 이름과(string), 요일별 먹을 메뉴(string), 그리고 먹지 못하는 메뉴(string[])가 저장
   #categories; // 요일별로 카테고리를 저장하는 배열. 0,1,2,3,4 -> 월화수목금 으로 보면 된다.
@@ -38,11 +26,7 @@ class App {
     this.#coaches = new Map();
     this.#categories = [];
     Console.print("점심 메뉴 추천을 시작합니다.");
-
-    for (let i = 0; i < 5; i++) {
-      const curNum = getCategories(this.#categories);
-      this.#categories.push(curNum);
-    }
+    this.#categories = randomCategory(this.#categories);
 
     console.log(this.#categories);
 
@@ -126,6 +110,7 @@ class App {
     }
 
     Console.print("추천을 완료했습니다.");
+    Console.close();
   }
 }
